+++ lib/cryptic/src/mem/memerror.rs
@@ -2,8 +2,8 @@ use cryptic_derive::Cryptee;
 
 #[derive(Cryptee)]
 pub enum MemError {
-    ///memory out of bounds: 0x{:02x} ; {:016b}
-    OutOfBounds(u16),
-    ///failed to read memory address: 0x{:02x} ; {:016b}
-    MemRead(u16),
+    ///memory out of bounds: 0x{:02x} ; {:032b}
+    OutOfBounds(u32),
+    ///failed to read memory address: 0x{:02x} ; {:032b}
+    MemRead(u32),
 }
